Extract shared function
