+++ packages/react/src/Breakout/Breakout.tsx
@@ -8,34 +8,13 @@ import { forwardRef } from 'react'
 import type { ForwardedRef } from 'react'
 import { BreakoutCell } from './BreakoutCell'
 import type { GridColumnNumber, GridPaddingSize, GridProps } from '../Grid/Grid'
+import { paddingClasses } from '../Grid/paddingClasses'
 
 export type BreakoutColumnNumber = GridColumnNumber
 export type BreakoutPaddingSize = GridPaddingSize
 export type BreakoutRowNumber = 1 | 2 | 3 | 4
 
 export type BreakoutProps = GridProps
-
-const paddingClasses = (
-  paddingBottom?: BreakoutPaddingSize,
-  paddingTop?: BreakoutPaddingSize,
-  paddingVertical?: BreakoutPaddingSize,
-): string[] => {
-  const classes = [] as string[]
-
-  if (paddingVertical) {
-    return [`ams-breakout--padding-vertical--${paddingVertical}`]
-  }
-
-  if (paddingBottom) {
-    classes.push(`ams-breakout--padding-bottom--${paddingBottom}`)
-  }
-
-  if (paddingTop) {
-    classes.push(`ams-breakout--padding-top--${paddingTop}`)
-  }
-
-  return classes
-}
 
 const BreakoutRoot = forwardRef(
   (
@@ -48,7 +27,7 @@ const BreakoutRoot = forwardRef(
       className={clsx(
         'ams-breakout',
         gapVertical && `ams-breakout--gap-vertical--${gapVertical}`,
-        paddingClasses(paddingBottom, paddingTop, paddingVertical),
+        paddingClasses('breakout', paddingBottom, paddingTop, paddingVertical),
         className,
       )}
     >
